+++ src/linspace.rs
@@ -19,7 +19,7 @@ pub type Linspace<F> = iter::Take<iter::Counter<F>>;
 ///            vec![0., 0.25, 0.5, 0.75, 1.0]);
 /// ```
 #[inline]
-pub fn linspace<F: Float>(a: F, b: F, n: uint) -> Linspace<F>
+pub fn linspace<F: Float>(a: F, b: F, n: usize) -> Linspace<F>
 {
     if n != 0 {
         let nf: F = NumCast::from(n).unwrap();
